reorganize componet file structure
