+++ soundscapes/app/frontend/SoundScapesFrontend/components/main_tab/main_tab.js
