+++ soundscapes/app/frontend/SoundScapesFrontend/reducers/root_reducer.js
@@ -1,6 +1,8 @@
 import { combineReducers } from 'redux';
-import entities from '../entities_reducer';
+import entities from './entities_reducer';
+import ui from './ui_reducer';
 
 export default combineReducers({
-  entities
+  entities,
+  ui
 });
